feat: added delete button to selected modal
